refactor: timeForToday 함수 utils.js로 이동
