+++ src/js/view/view.js
@@ -1,4 +1,5 @@
 import { $ } from "../utils/DOM.js";
+import { timeForToday } from "../utils/utils.js";
 
 const $videoList = $(".video-list");
 
@@ -28,27 +29,4 @@ export const paintVideoList = (data) => {
 
     $videoList.innerHTML += searchResultVideoListTemplate(imageUrl, title, channelName, publishedDate);
   });
-};
-
-const timeForToday = (value) => {
-  const today = new Date();
-  const timeValue = new Date(value);
-
-  const betweenTime = Math.floor((today.getTime() - timeValue.getTime()) / 1000 / 60);
-  if (betweenTime < 1) return "방금전";
-  if (betweenTime < 60) {
-    return `${betweenTime}분전`;
-  }
-
-  const betweenTimeHour = Math.floor(betweenTime / 60);
-  if (betweenTimeHour < 24) {
-    return `${betweenTimeHour}시간전`;
-  }
-
-  const betweenTimeDay = Math.floor(betweenTime / 60 / 24);
-  if (betweenTimeDay < 365) {
-    return `${betweenTimeDay}일전`;
-  }
-
-  return `${Math.floor(betweenTimeDay / 365)}년전`;
 };
